requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements/dev.txt --resolver=backtracking pyproject.toml
#
astroid==2.15.6
    # via pylint
attrs==23.1.0
    # via big-salad (pyproject.toml)
black==23.7.0
    # via big-salad (pyproject.toml)
boto3==1.28.2
    # via big-salad (pyproject.toml)
botocore==1.31.2
    # via
    #   boto3
    #   s3transfer
certifi==2023.5.7
    # via requests
charset-normalizer==3.2.0
    # via requests
click==8.1.5
    # via
    #   big-salad (pyproject.toml)
    #   black
coverage[toml]==7.2.7
    # via
    #   big-salad (pyproject.toml)
    #   pytest-cov
deepdiff==6.3.1
    # via big-salad (pyproject.toml)
dill==0.3.6
    # via pylint
docker==6.1.3
    # via big-salad (pyproject.toml)
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via pylint
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lazy-object-proxy==1.9.0
    # via astroid
mccabe==0.7.0
    # via pylint
mypy-extensions==1.0.0
    # via black
ordered-set==4.1.0
    # via deepdiff
orjson==3.9.2
    # via big-salad (pyproject.toml)
packaging==23.1
    # via
    #   black
    #   docker
    #   pytest
pathspec==0.11.1
    # via black
platformdirs==3.8.1
    # via
    #   black
    #   pylint
pluggy==1.2.0
    # via pytest
podman==4.5.1
    # via big-salad (pyproject.toml)
psycopg2-binary==2.9.6
    # via big-salad (pyproject.toml)
pylint==2.17.4
    # via big-salad (pyproject.toml)
pymysql==1.1.0
    # via big-salad (pyproject.toml)
pytest==7.4.0
    # via
    #   big-salad (pyproject.toml)
    #   pytest-cov
pytest-cov==4.1.0
    # via big-salad (pyproject.toml)
python-dateutil==2.8.2
    # via botocore
pyxdg==0.28
    # via podman
pyyaml==6.0
    # via big-salad (pyproject.toml)
requests==2.31.0
    # via
    #   docker
    #   podman
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.18
    # via big-salad (pyproject.toml)
tomlkit==0.11.8
    # via pylint
typing-extensions==4.7.1
    # via sqlalchemy
urllib3==1.26.16
    # via
    #   botocore
    #   docker
    #   podman
    #   requests
websocket-client==1.6.1
    # via docker
wrapt==1.15.0
    # via astroid
